changing number of layers/neurons/filters in each layer during training process
The mclfy project https://github.com/NLeSC/mcfly does an optimization over parameters such as the amount of layers. It was written for time series classification but the concept can be adapted for other use cases. It creates a set of possible architectures, trains them with a subset of the data, and compares the results. The most promising architecture is then trained on the whole data. Stuff like which architecture (Conv or DeepConvLSTM), how many layers within these etc., are considered variables for which it optimizes.