forcing garbage collection
If you create a large object and delete it again, Python has probably released the memory, but the memory allocators involved don’t necessarily return the memory to the operating system, so it may look as if the Python process uses a lot more virtual memory than it actually uses.
The only really reliable way to ensure that a large but temporary use of memory DOES return all resources to the system when it's done, is to have that use happen in a subprocess, which does the memory-hungry work then terminates.

https://pymotw.com/2/gc/
https://pymotw.com/3/gc/
https://www.youtube.com/watch?v=hg01TTzi6Xk
	
Python will automatically collect objects with circular references, unless any of the objects in a reference cycle have del methods. If that's the case, garbage objects are moved to the gc.garbage list, and you will have to manually break the reference cycles. It's better to try to avoid having both del methods and reference cycles.