Bayesian
This program uses the pandas library to read the CSV file and the pgmpy library to learn the Bayesian Network. The BayesianEstimator class from pgmpy is used to estimate the structure of the network based on the data in the columns of the DataFrame. The estimate method is used to learn the structure of the network, and the edges method is used to print the structure of the learned network.

You can customize the program by changing the parameters of the BayesianEstimator and estimate functions and by adding additional code to process or analyze the learned network.