LDA
Latent Dirichlet allocation (LDA) is a method for unsupervised classification of documents into natural groups based on their content 1. It is a particularly popular method for fitting a topic model. It treats each document as a mixture of topics, and each topic as a mixture of words. This allows documents to “overlap” each other in terms of content, rather than being separated into discrete groups, in a way that mirrors typical use of natural language.

To perform LDA on a text, you can use the topicmodels package in R. The package takes a Document-Term Matrix as input and produces a model that can be manipulated and visualized with dplyr and ggplot.