tool for working with executable papers
As of 2024, this project is archived and unmaintained. While is has achieved its mission of demonstrating that unifying computational reproducibility and provenance tracking is doable and useful, it has also demonstrated that Python is not a suitable platform to build on for reproducible research. Breaking changes at all layers of the software stack are too frequent.”
https://github.com/activepapers/activepapers-python